added an update function
You can now toggle whether you want the script to check for a newer version or not. It is possible to include or exclude beta release versions.
A backup from the old version will be created as well.
I also changed the way of saving settings in the main GUI. You no longer need to close the window because a simple right-click will do the trick.