improved auto update and fixed issues
The PowerShell file responsible for updating to a higher version will now move every file instead of copying them. This makes room for the new version of the script to "replace" them with the exception of the config file.
This had to be done because the new script won't replace existing files.
I also correctly suspended every hotkey until GTA is launched to avoid annoying side effects.
Fixed a bug which caused the script to display a MsgBox at the wrong time.
Some functions have been moved to different lines.