Development (#4)
* workflow patch

* workflow patch

* added AutoHotkey compiler for CI workflows

* workflow patch

* workflow patch

* workflow patch

* workflow patch

* workflow patch

* workflow patch

* workflow patch

* workflow patch

* workflow patch

* workflow patch

* workflow patch

* workflow patch

* workflow patch

* workflow patch

* workflow patch

* added a README file

This README file contains some information about this project.

* changed executable properties

* added first implementation of recording macros

You can now record the macros for the two cash deposit hotkeys. The README file has been improved with a macro FAQ.

* updated the README file

* attempt to fix false virus alert

I suspect the recently added AutoHotkey32.EXE is the problem here. For now I tried to put it into a ZIP folder.

* added an update function

You can now toggle whether you want the script to check for a newer version or not. It is possible to include or exclude beta release versions.
A backup from the old version will be created as well.
I also changed the way of saving settings in the main GUI. You no longer need to close the window because a simple right-click will do the trick.

* workflow patch

* fixed issue which caused missing files

I added extra quotation marks to the PowerShell command that extracts the ZIP files because when the path contained empty spaces, the extraction would fail.

* added a confirmation prompt when launching the first time

You will now have to confirm the creation of the script's files.

* improved auto update and fixed issues

The PowerShell file responsible for updating to a higher version will now move every file instead of copying them. This makes room for the new version of the script to "replace" them with the exception of the config file.
This had to be done because the new script won't replace existing files.
I also correctly suspended every hotkey until GTA is launched to avoid annoying side effects.
Fixed a bug which caused the script to display a MsgBox at the wrong time.
Some functions have been moved to different lines.

* further improved auto update

The script will from now on be launched from the temp directory. This avoids file operation errors because the script cannot move itself while it is running.
I also improved the clarity of the code.

* improved auto update

I improved the code readability and structure.
The version is now no longer a fixed value inside the code. It can be set via a text file. This means the GitHub workflow is able to change the version automatically.

* workflow patch

* workflow patch