expanded main GUI open file menu
You can now open the script directory and the support file directory from the main GUI.
It is also now possible for the script to load hotkeys and create an array with the corresponding custom macro objects.
But there is still no GUI integration for creating and managing those objects yet.