added auto import feature for macros
When you copy macro files together with the macro config file, the script will attempt to update the macro file path in that copied macro config file for the new hotkeys.
In other words, if you have a hotkey called "one" and it's macro is saved under "C:\A\GTAV_Tweaks\macros\one.ahk" and you copy the content of that macro folder to "C:\B\GTAV_Tweaks\macros",
the macro file path for the hotkey "one" will automatically change from "C:\A\GTAV_Tweaks\macros\one.ahk" to "C:\B\GTAV_Tweaks\macros\one.ahk"

This does only work, when the files are not renamed after moving them and if the macro file is located in the folder "macros" inside the "GTAV_Tweaks" folder.