added German language support
I created the included Languages.ahk file, which contains all text box values in English and in German. You can change the language in the config file or in the main window.
Adding other languages should not be that hard. Just copy the English framework and change the string values accordingly.