added German built-in hotkey description
The two built-in hotkeys now have a German translation. Which text will be saved for them depends on the the selected language when they were created.
It is possible to update their description language by simply deleting them. The will be recreated automatically with the currently selected language.
I also renamed the language option "AUTO" to "SYSTEM" which will attempt to use the system language.