added option to force updates
You can now force an update to the latest version depending on the update settings. This is intended to update clients that already have a beta version installed. The problem is, that the auto update script cannot determine if the beta version on GitHub received an update or not. Maybe I will have to change the versioning for beta versions to allow the auto update script to see the updates.