fixed a force update bug and translated TrayTip messages
I fixed a bug that caused the forced update option to fail, because there was a check if the script was already on the latest version. If this check was true, the update was canceled, even tough the function was called with the force update parameter set to true.
The TrayTip messages are now available in German.