saving color detection progress
The color recognition works fine on normal windows apps, but it does not work inside the GTA V browser. For some reason the color inside the GTA V browser is changed to white after a few milliseconds of not moving the mouse cursor. For now I will just save my work and continue on other aspects of improving the macro recorder.