further integrated interactive tutorials
There is now a class called "InteractiveTutorial" which creates a little GUI capable of displaying text and allowing the user to navigate through a series of predefined steps.
I also changed the main GUI so that clicking on the help menu will open the help GUI immediately.
There is also a small tutorial for the help GUI available in German and English.