improved autostart
The autostart script will now check for other instances from other instances of GTAV Tweaks and stop them. This should avoid multiple instances of GTAV Tweaks starting together with GTA V.

When updating the script, the autostart script will be stopped to avoid an error because the file is already used and cannot be copied or moved.