fixed an update bug
There was a bug that caused the script to launch the old version during the update process. This was caused by the PowerShell instances that usually launches GTAV Tweaks when it detects a running instance of GTA V.

Due to this remnant of the old script not being deleted during the update, the script would attempt to restart GTAV Tweaks and cause problems during the restart.

I also added a bunch of $null to avoid future errors in all PowerShell scripts.