improved workflow reliability
Some Write-Host statements blasted their outputs straight into the stdOut. I fixed that with a bunch of $null = ...
I also removed the automatic generated release notes, as they are wrong most of the time because they are comparing the current version with an already outdated one instead of the previous version.