fixed auto update date issue
I initially thought that formatting all dates into a specific format would resolve issues. It did, but only when the system the script was started on had the correct regional settings.

On U.S. systems (and a few other regions) the forced format could not be interpreted, causing the update process to fail.

Forcing an international standard (ISO 8601) should help preventing this issue in the future.