changed automatic releases
The currentVersion file has not been updated when packing it into the .MSI installer. This has been fixed in the workflow.
The update-beta-release workflow will now always deliver a new beta release version instead of keeping the old one.

For example:

v0.4.0.0-beta will become v0.4.0.1-beta instead of staying at v0.4.0.0-beta

The old beta release and tag will be deleted.

This change was necessary because the .MSI installer always needs to change the revision part of the version when anything changes to allow the new installer to update existing versions.