implemented update user interface
The PowerShell update script has been modified to allow for the new installation process involving a simple .MSI file instead of the GTAV_Tweaks.zip folder.
It now has an option to install the .MSI file automatically or allows the user to configure the installation, if they choose to do so.
The update script no longer enables the GTAV Tweaks tutorial after changing the version.