completely removed automatic update
It is sad but I think this is the only possible way. I don't plan on buying a digital software signing certificate to sign my program but it keeps getting flagged as malware by Windows Defender to the point where you can't even download it anymore without trouble.
To make the application "less suspicious" I decided to remove the parts of the update PowerShell script that automatically download and install the .MSI installer.
From now on GTAV Tweaks will download the new installer file and tell the user how to proceed from there.